requirements.txt
fastapi
uvicorn
elasticsearch
redis
psycopg2-binary
newspaper3k
pydantic
pydantic-settings
pytest
pytest-asyncio
httpx